requirements.txt
fastapi
uvicorn
pandas
openpyxl
requests
python-dotenv
grpcio
grpcio-tools
protobuf
pandas
openpyxl